Mejora UI con logica
